For #7067: Add top site widget
