+++ Client/Frontend/photon-colors.swift
@@ -6,6 +6,9 @@
  From https://github.com/FirefoxUX/photon-colors/#readme */
 import UIKit
 
+// Used as backgrounds for favicons
+public let DefaultFaviconBackgroundColors = ["2e761a", "399320", "40a624", "57bd35", "70cf5b", "90e07f", "b1eea5", "881606", "aa1b08", "c21f09", "d92215", "ee4b36", "f67964", "ffa792", "025295", "0568ba", "0675d3", "0996f8", "2ea3ff", "61b4ff", "95cdff", "00736f", "01908b", "01a39d", "01bdad", "27d9d2", "58e7e6", "89f4f5", "c84510", "e35b0f", "f77100", "ff9216", "ffad2e", "ffc446", "ffdf81", "911a2e", "b7223b", "cf2743", "ea385e", "fa526e", "ff7a8d", "ffa7b3" ]
+
 extension UIColor {
     struct Photon {
         static let Magenta50 = UIColor(rgb: 0xff1ad9)
